Update patch set 4
Patch Set 4:

Hi,

In the solution test section we need to explicitly call out about the LR scaling (# of LRs that will be configured in the system) since this change should not impact the existing scale numbers of LR/VN (or any other objects) etc. Only new variable that is being introduced and scale tested as part of this change is the no. of routes received from PNF that we decided to qualify to a average 25 and max of not > 100 per LR.

Patch-set: 4
Reviewer: Gerrit User 288 <288@c72cd27d-1743-4a55-84a4-cd13b8400c52>